requirements.txt
paho-mqtt==1.6.1
pandas==1.4.2
tabulate==0.9.0